Fix: rename model_versions to model
