+++ tmaze/src/settings/model/ver_2.rs
